requirements.txt
pip==18.0

# Core
pytz==2018.4
Django==2.1.1
django-configurations==2.1
gunicorn==19.8.1
newrelic==3.2.2.94

# For the persistence stores
psycopg2-binary==2.7.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest API
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==2.0
django-rest-swagger>=2.2.0
django-cors-headers==2.4.0
coreapi>=2.3.3

# Developer Tools
ipdb==0.11
ipython==6.4.0
mkdocs==0.17.4
flake8==3.5.0
django-debug-toolbar>=1.9.1

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1
pytest-django


# Static and Media Storage
django-storages==1.6.6
boto3==1.7.39

biopython>=1.72
ipykernel>=4.9.0
pandas>=0.23.4
numpy>=0.15.1
scipy>=1.1.0
matplotlib>=2.2.3

jsonschema>=2.6.0
django-extra-fields>=1.0.0
PyPDF2>=1.26.0
watchdog>=0.9.0
requests>=2.19.1
coloredlogs>=10.0

